+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v11.0.7
+* #1857 - Add LibCustomGlow animations.
 * Fix: #2215 - Spell Charges condition not updating for countable spells without true charges.
 * Fix: #2217 - Error in item cooldown conditions
 * Fix: #2220 - GCD state not ending when GCD ends.